start to create Profile settings

+++ src/App.js
@@ -3,6 +3,7 @@ import Menu from "./Menu/Menu";
 import SearchPanel from "./SearchPanel/SearchPanel";
 import HorizontalMenu from "./HorizontalMenu/HorizontalMenu";
 import NoteBoard from "./NoteBoard/NoteBoard";
+import ProfileSettings from "./ProfileSettings/ProfileSettings";
 
 export default class App extends Component {
   constructor(props) {
@@ -13,16 +14,20 @@ export default class App extends Component {
           id: "12312sasda",
           completed: false,
           title: "Example note",
-          text: "Write some note. To do this, in the menu, click 'new note'",
+          text: "Write some note. To do this click 'new note'",
           isCompleted: false
         }
       ],
-      modal: false
+      modal: false,
+      profSettings: false,
+      homePage: true
     };
     this.addItem = this.addItem.bind(this);
     this.deleteItem = this.deleteItem.bind(this);
     this.onEnterCloseModal = this.onEnterCloseModal.bind(this);
-    this.completedToggle = this.completedToggle.bind(this);
+    /* this.completedToggle = this.completedToggle.bind(this); */
+    this.openProfSettings = this.openProfSettings.bind(this);
+    this.openHomePage = this.openHomePage.bind(this);
   }
 
   getLocalStorage() {
@@ -86,7 +91,32 @@ export default class App extends Component {
     });
   }
 
-  completedToggle(id) {
+  /* openProfSettings() {
+    this.setState({
+      profSettings: true,
+      homePage: false
+    });
+  } */
+
+  openProfSettings = () => {
+    if (!this.state.profSettings) {
+      this.setState({
+        profSettings: true,
+        homePage: false
+      });
+    }
+  };
+
+  openHomePage = () => {
+    if (!this.state.homePage) {
+      this.setState({
+        homePage: true,
+        profSettings: false
+      });
+    }
+  };
+
+  /*   completedToggle(id) {
     this.setState(({ todos }) => {
       const index = todos.findIndex(elem => elem.id === id);
 
@@ -103,32 +133,60 @@ export default class App extends Component {
         todos: newArr
       };
     });
-  }
+  } */
 
   render() {
-    const { todos, modal } = this.state;
+    const { todos, modal, profSettings, homePage } = this.state;
     const { isCompleted } = this.state.todos;
-
-    return (
-      <div className="App">
-        <div className="container">
-          <Menu />
-          <div className="notes-block-container">
-            <SearchPanel />
-            <HorizontalMenu
-              modal={modal}
-              addItem={this.addItem}
-              onEnterCloseModal={this.onEnterCloseModal}
+    if (profSettings) {
+      return (
+        <div className="App">
+          <div className="container">
+            <Menu
+              openProfSettings={this.openProfSettings}
+              openHomePage={this.openHomePage}
             />
-            <NoteBoard
-              todos={todos}
-              onDelete={this.deleteItem}
-              onCompleted={this.completedToggle}
-              isCompleted={isCompleted}
-            />
+            <div className="notes-block-container">
+              <SearchPanel />
+              <HorizontalMenu
+                modal={modal}
+                addItem={this.addItem}
+                onEnterCloseModal={this.onEnterCloseModal}
+                isHomePage={homePage}
+                openHomePage={this.openHomePage}
+              />
+              <ProfileSettings />
+            </div>
           </div>
         </div>
-      </div>
-    );
+      );
+    } else if (homePage) {
+      return (
+        <div className="App">
+          <div className="container">
+            <Menu
+              openProfSettings={this.openProfSettings}
+              openHomePage={this.openHomePage}
+            />
+            <div className="notes-block-container">
+              <SearchPanel />
+              <HorizontalMenu
+                modal={modal}
+                addItem={this.addItem}
+                onEnterCloseModal={this.onEnterCloseModal}
+                isHomePage={homePage}
+                openHomePage={this.openHomePage}
+              />
+              <NoteBoard
+                todos={todos}
+                onDelete={this.deleteItem}
+                onCompleted={this.completedToggle}
+                isCompleted={isCompleted}
+              />
+            </div>
+          </div>
+        </div>
+      );
+    }
   }
 }
